Parallelise the for-loop which calculates the harmonics of the signal
Previously, the calculation of harmonics for the signal was run in parallel alongside the surrogates; however, this did not provide a performance improvement when surrogates were not used.

Now, the for loop which calculates the harmonics has been parallelised. When the number of surrogates is large, this for loop will not be parallelised to reduce overhead.